refactor: move referenceLinkUtils to src/utils/referenceLink
