+++ src/layouts/ReferenceItemLayout.astro
@@ -24,7 +24,7 @@ import _ from 'lodash';
 import { Code } from 'astro:components';
 import OutdatedTranslationBanner from "@components/OutdatedTranslationBanner/index.astro";
 import { checkTranslationBanner } from "../utils/translationBanner";
-import { normalizeP5ReferenceLinks } from "./referenceLinkUtils";
+import { normalizeP5ReferenceLinks } from "../utils/referenceLink";
 
 const { entry, relatedEntries } = Astro.props;
 const currentLocale = getCurrentLocale(Astro.url.pathname);
