+++ tsconfig.json
@@ -28,6 +28,7 @@
       "@tests/*": ["./src/tests/*"],
       "@types/*": ["./src/types/*"],
       "@i18n/*": ["./src/i18n/*"],
+      "@utils/*": ["./src/utils/*"],
       "@styles/*": ["./styles/*"],
     }
   }
